requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.1
jinja2==3.1.4
python-multipart==0.0.12
sqlalchemy==2.0.36
psycopg2-binary==2.9.10
requests>=2.32.0
pdfplumber>=0.11.4
beautifulsoup4>=4.12.3
apscheduler>=3.10.4
python-dotenv>=1.0.1
pytest>=8.3.0
httpx>=0.27.0
